% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_example.R
\name{mol_example}
\alias{mol_example}
\title{Get path to example .mol file}
\usage{
mol_example(path = NULL)
}
\arguments{
\item{path}{Name of file.  If \code{NULL} (default), the options for example .mol
files will be listed instead.}
}
\description{
\code{volcalc} comes bundled with some example .mol files in its \code{inst/extdata}
directory.  This function provides easy access to them.
}
\details{
File names are the KEGG identifiers.  Compound names are as follows:
\itemize{
\item C00031: D-Glucose
\item C08491: (-)-Jasmonic acid
\item C16181: beta-2,3,4,5,6-Pentachlorocyclohexanol
\item C16286: Geosmin
\item C16521: Isoprene
}
}
\examples{
#list all examples
mol_example()

#return path to specific example file
mol_example("C16181.mol")
}
