% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_propOB.R
\name{safe_propOB}
\alias{safe_propOB}
\alias{ob_problems}
\title{"Safe" version of ChemmineR::propOB()}
\usage{
safe_propOB(sdfSet, ob_err_pass = FALSE)

ob_problems(x)
}
\arguments{
\item{sdfSet}{an \code{sdfSet} object}

\item{ob_err_pass}{logical; return results even if there are warnings or
errors? Defaults to \code{FALSE}}

\item{x}{data frame created by \code{\link[=safe_propOB]{safe_propOB()}}}
}
\value{
A data frame with any messages from OpenBabel stored as the
attribute \code{"OB_stderr"}.  See \code{\link[ChemmineR:propOB]{ChemmineR::propOB()}} for additional details.
}
\description{
A "safe" version of \code{\link[ChemmineR:propOB]{ChemmineR::propOB()}} that returns \code{NA}s when OpenBabel
produces warnings or errors.
}
\details{
\code{\link[ChemmineR:propOB]{ChemmineR::propOB()}} can result in errors from OpenBabel that are difficult
to capture because they are not R errors or warnings (i.e. they are not
captured by \code{\link[=capture.output]{capture.output()}} or \code{\link[=sink]{sink()}}). This solution (thanks to Jeron
Ooms) runs \code{\link[ChemmineR:propOB]{ChemmineR::propOB()}} in a separate R session and captures
\code{stderr} in a temp file. By default, if there is anything at all in \code{stderr},
all values returned are \code{NA}s, assuming that the results are not reliable.
This behavior can be overridden with \code{ob_err_pass}, but it is a good idea to
inspect warnings and errors with \code{\link[=ob_problems]{ob_problems()}} and use resulting data at
your own risk!
}
\examples{
\dontrun{
tmp <- tempdir()
sdf <- get_mol_kegg("C00157", dir = tmp)
sdf <- ChemmineR::read.SDFset(file.path(tmp, "C00157.mol"))
prop <- ChemmineR::propOB(sdf)
prop_safe <- safe_propOB(sdf)
ob_problems(prop_safe)
}
}
\seealso{
\code{\link[=ob_problems]{ob_problems()}}
}
