% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_compound_mol.R
\name{save_compound_mol}
\alias{save_compound_mol}
\title{Download compound .mol file}
\usage{
save_compound_mol(
  compound_id = NULL,
  compound_formula = NULL,
  pathway_id = NULL,
  path = "data",
  redownload = FALSE
)
}
\arguments{
\item{compound_id}{character string that is 5 digits prepended with a "C"}

\item{compound_formula}{character string of compound formula}

\item{pathway_id}{optional character string specifying KEGG pathway ID, in format of 5 digits prepended with "map"}

\item{path}{optional parameter to set relative path to location to download data with default of creating "data" folder in home directory}

\item{redownload}{download file again even if it has already been downloaded at path}
}
\value{
downloaded .mol file for compound in path folder
}
\description{
Using KEGG ID values for compound and pathway, download
and clean up corresponding mol file
}
