% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcalc-package.R
\docType{package}
\name{volcalc-package}
\alias{volcalc}
\alias{volcalc-package}
\title{volcalc: Calculate Volatility of Chemical Compounds}
\description{
Use this package to calculate estimated volatility values for individual compounds of interest or for all compounds in a pathway. Calculation uses the SIMPOL method (Prankow and Asher, 2008), and is only currently applicable to compounds and pathways in KEGG.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://meredith-lab.github.io/volcalc/}
}

}
\author{
\strong{Maintainer}: Kristina Riemer \email{kristinariemer@arizona.edu} (\href{https://orcid.org/0000-0003-3802-3331}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eric R. Scott (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})
}

}
\keyword{internal}
