% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pathway_vol.R
\name{calc_pathway_vol}
\alias{calc_pathway_vol}
\title{Calculate volatility estimates for all compounds in a pathway}
\usage{
calc_pathway_vol(pathway_id, path)
}
\arguments{
\item{pathway_id}{character string that is 5 digits prepended with "map"}

\item{path}{relative path to location to download data}
}
\value{
dataframe with one row per compound and columns for basic compound info, functional group counts, and volatility estimate and category
}
\description{
Downloads all compound .mol files associated with a given KEGG pathway and
counts functional groups and calculates volatility value and category for all
}
