% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vol.R
\name{calc_vol}
\alias{calc_vol}
\title{Calculate volatility estimate for a compound}
\usage{
calc_vol(
  input,
  from = c("mol_path"),
  method = c("simpol1"),
  return_fx_groups = FALSE,
  return_calc_steps = FALSE
)
}
\arguments{
\item{input}{a path to a .mol file or a character representation such as
SMILES or InChI}

\item{from}{the form of \code{input} (currently only a path to a .mol file is
implemented)}

\item{method}{the method for calculating estimated volatility. Currently only
the SIMPOL.1 method is implemented---see \code{\link[=simpol1]{simpol1()}} for more details.}

\item{return_fx_groups}{When \code{TRUE}, includes functional group counts in
final dataframe.}

\item{return_calc_steps}{When \code{TRUE}, includes intermediate volatility
calculation steps in final dataframe.}
}
\value{
a tibble with columns of basic compound info and volatility value
and category created by \code{\link[=get_fx_groups]{get_fx_groups()}} and \code{\link[=simpol1]{simpol1()}}.
}
\description{
Relative volatility value and category is estimated for specified compound
using group contribution methods.
}
\seealso{
\code{\link[=get_fx_groups]{get_fx_groups()}}, \code{\link[=simpol1]{simpol1()}}
}
