% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vol.R
\name{calc_vol}
\alias{calc_vol}
\title{Calculate volatility estimate for compound}
\usage{
calc_vol(
  pathway_id,
  path,
  compound_id = NULL,
  compound_formula = NULL,
  redownload = FALSE,
  save_file = TRUE,
  get_groups = TRUE,
  fx_groups_df = NULL
)
}
\arguments{
\item{pathway_id}{character string that is 5 digits prepended with "map"}

\item{path}{relative path to location to download data}

\item{compound_id}{character string that is 5 digits prepended with a "C"}

\item{compound_formula}{character string of compound formula}

\item{redownload}{download file again even if it has already been downloaded at path}

\item{save_file}{save compound mol file using save_compound_mol function}

\item{get_groups}{get dataframe of compound functional groups using get_fx_groups function}

\item{fx_groups_df}{dataframe of functional group counts for compounds, optional if reading functional groups dataframe in directly}
}
\value{
input dataframe with new columns for volatility value and category
}
\description{
Using functional group counts from get_fx_groups(), volatility is estimated
using the SIMPOL formula
}
