% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vol.R
\name{calc_vol}
\alias{calc_vol}
\title{Calculate volatility estimate for compound}
\usage{
calc_vol(fx_groups_df)
}
\arguments{
\item{fx_groups_df}{dataframe of functional group counts for compounds}
}
\value{
input dataframe with new columns for volatility value and category
}
\description{
Using functional group counts from get_fx_groups(), volatility is estimated
using the SIMPOL formula
}
