% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vol.R
\name{calc_vol}
\alias{calc_vol}
\title{Calculate volatility estimate for compound}
\usage{
calc_vol(
  input,
  from = c("mol_path"),
  return_fx_groups = FALSE,
  return_calc_steps = FALSE
)
}
\arguments{
\item{input}{a path to a .mol file or a character representation such as SMILES or InChI}

\item{from}{the form of \code{input} (currently only a path to a .mol file is implemented)}

\item{return_fx_groups}{When \code{TRUE}, includes functional group counts in
final dataframe.}

\item{return_calc_steps}{When \code{TRUE}, includes intermediate volatility
calculation steps in final dataframe.}
}
\value{
Dataframe with columns of basic compound info and volatility value
and category. See documentation for column descriptions.
}
\description{
Volatility value and category is estimated for specified compound using the
SIMPOL formula
}
