% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vol.R
\name{calc_vol}
\alias{calc_vol}
\title{Calculate volatility estimate for a compound}
\usage{
calc_vol(
  input,
  from = c("mol_path", "smiles"),
  method = c("meredith", "simpol1"),
  temp_c = 20,
  environment = c("clean", "polluted", "soil"),
  validate = TRUE,
  return_fx_groups = FALSE,
  return_calc_steps = FALSE
)
}
\arguments{
\item{input}{A path to a .mol file or a SMILES string.}

\item{from}{The form of \code{input}. Either \code{"mol_path"} (default) or \code{"smiles"}.}

\item{method}{The method for calculating estimated volatility. See
\code{\link[=simpol1]{simpol1()}} for more details.}

\item{temp_c}{numeric; For methods that allow it, specify temperature (in
degrees C) at which log10(P) estimates are calculated.}

\item{environment}{The environment for calculating relative volatility
categories. RVI thresholds for low, moderate, and high volatility are as
follows: \code{"clean"} (clean atmosphere, default) -2, 0, 2; \code{"polluted"}
(polluted atmosphere) 0, 2, 4; \code{"soil"} 4, 6, 8. For more information about
these thresholds see Meredith et al. (2023) and Donahue et al. (2006).}

\item{validate}{logical; if \code{TRUE} (default), results are checked for
possible errors in parsing by Open Babel and \code{NA}s are returned if possible
errors are found. Setting to \code{FALSE} bypasses these checks—use at your own
risk! Validation is not available on Windows. See \strong{Details} of
\code{\link[=get_fx_groups]{get_fx_groups()}} for more information.}

\item{return_fx_groups}{When \code{TRUE}, the returned tibble includes functional
group counts.}

\item{return_calc_steps}{When \code{TRUE}, the returned tibble includes
intermediate volatility calculations. See \strong{Details}.}
}
\value{
A tibble with relative volatility index (\code{rvi}) and volatility
category (\code{category}).
}
\description{
Estimate relative volatility index value and category for specified compounds
using group contribution methods.
}
\details{
\eqn{\textrm{log}_{10}C^\ast} is used for the calculated relative
volatility index (\code{rvi}). \eqn{\textrm{log}_{10}C^\ast =
  \textrm{log}_{10}(PM/RT)} where \eqn{P} is the estimated vapor pressure for
the compound, \eqn{M} is molecular mass of the compound, \eqn{R} is the
universal gas constant, and \eqn{T} is temperature (293.14K or 20ºC).  When
\code{return_calc_steps = TRUE}, the log of estimated vapor pressure, \code{log10_P},
and \eqn{\textrm{log}_{10}(M/RT)}, \code{log_alpha}, are also returned.
}
\examples{
\donttest{
mol_paths <- mol_example()
calc_vol(mol_paths)

# Return functional group counts from get_fx_groups()
calc_vol(mol_paths,  return_fx_groups = TRUE)

# Return intermediate calculations
calc_vol(mol_paths, return_calc_steps = TRUE)
}
}
\references{
Donahue, N.M., Robinson, A.L., Stanier, C.O., Pandis, S.N., 2006.
Coupled Partitioning, Dilution, and Chemical Aging of Semivolatile Organics.
Environ. Sci. Technol. 40, 2635–2643. \doi{10.1021/es052297c}

Meredith L, Ledford S, Riemer K, Geffre P, Graves K, Honeker L, LeBauer D,
Tfaily M, Krechmer J. 2023. Automating methods for estimating metabolite
volatility. Frontiers in Microbiology. \doi{10.3389/fmicb.2023.1267234}
}
\seealso{
\code{\link[=get_fx_groups]{get_fx_groups()}}, \code{\link[=simpol1]{simpol1()}}
}
