% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fx_groups.R
\name{get_fx_groups}
\alias{get_fx_groups}
\title{Count functional groups of a compound}
\usage{
get_fx_groups(compound_sdf)
}
\arguments{
\item{compound_sdf}{object created by \code{ChemmineR::read.SDFset()}}
}
\value{
Dataframe with columns of basic compound info and functional group
counts.
}
\description{
Return functional group counts relevant to volatility calculation for
specified compound
}
\note{
Users do not typically need to interact with this function, but it is
used internally by \code{calc_vol()}
}
\examples{
get_df <- get_mol_kegg(compound_ids = "C16181", dir = tempdir())
mol <- ChemmineR::read.SDFset(get_df$mol_path)
get_fx_groups(mol)

}
