% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fx_groups.R
\name{get_fx_groups}
\alias{get_fx_groups}
\title{Count functional groups of a compound}
\usage{
get_fx_groups(compound_sdf)
}
\arguments{
\item{compound_sdf}{object created by \code{ChemmineR::read.SDFset()}}
}
\value{
a tibble with columns of basic compound info and functional group
counts.
}
\description{
Return functional group counts relevant to calculating estimated volatility
for specified compounds. Users will not typically interact with this function
directly, but rather by using \code{calc_vol()}.
}
\note{
This function currently does not capture hydroperoxide, nitrophenol,
nitroesther, alicyclic ether, aromatic ether, or armoatic amine groups.
Contributions of SMARTS strings to capture these groups are welcome.
}
\examples{
mol_path <- mol_example("C16181.mol")
sdf <- ChemmineR::read.SDFset(mol_path)
fx_groups <- get_fx_groups(sdf)

}
