% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fx_groups.R
\name{get_fx_groups}
\alias{get_fx_groups}
\title{Count functional groups of a compound}
\usage{
get_fx_groups(compound_id, pathway_id = NULL, path = "data")
}
\arguments{
\item{compound_id}{character string that is 5 digits prepended with a "C"}

\item{pathway_id}{optional character string specifying KEGG pathway ID, in format of 5 digits prepended with "map"}

\item{path}{optional parameter to set relative path to location to download data with default of creating "data" folder in home directory}
}
\value{
single row dataframe with columns for numbers of different functional groups and basic compound details
}
\description{
Using a mol file for compound of interest, return number of
many functional groups in a dataframe
}
