% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fx_groups.R
\name{get_fx_groups}
\alias{get_fx_groups}
\title{Count compound functional groups}
\usage{
get_fx_groups(compound_sdf)
}
\arguments{
\item{compound_sdf}{a \link[ChemmineR:SDFset-class]{ChemmineR::SDFset} object returned by
\code{\link[ChemmineR:read.SDFset]{ChemmineR::read.SDFset()}} or \code{\link[ChemmineR:smiles2sdf]{ChemmineR::smiles2sdf()}}, for example.}
}
\value{
a tibble with columns of basic compound info and functional group
counts.
}
\description{
Returns functional group counts relevant to calculating estimated volatility
for specified compounds. Users will not typically interact with this function
directly, but rather by using \code{\link[=calc_vol]{calc_vol()}}.
}
\note{
This function currently does \strong{not} capture the following functional
groups used in SIMPOL.1:
\itemize{
\item carbon number on the acid-side of amide
\item C=C-C=O in a non-aromatic ring
\item alicyclic ether
\item aromatic ether
\item aromatic amine
\item carbonylperoxynitrate
\item hydroperoxide
\item carbonylperoxyacid
\item nitrophenol,
\item nitroesther
}

Contributions of SMARTS strings to capture these groups are welcome.
}
\examples{
mol_path <- mol_example()[1]
sdf <- ChemmineR::read.SDFset(mol_path)
fx_groups <- get_fx_groups(sdf)

}
\seealso{
\code{\link[=calc_vol]{calc_vol()}}
}
