% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mol_kegg.R
\name{get_mol_kegg}
\alias{get_mol_kegg}
\title{Download compound .mol files from KEGG}
\usage{
get_mol_kegg(compound_ids, pathway_ids, dir, force = FALSE)
}
\arguments{
\item{compound_ids}{Character vector of KEGG compound IDs—5 digits
prepended with a "C".}

\item{pathway_ids}{Character vector of KEGG pathway or pathway module IDs—5
digits prepended with "map" or "M", respectively.}

\item{dir}{Path to a folder to save .mol files in. Folder will be created if
it does not already exist.}

\item{force}{Logical; by default (\code{FALSE}), .mol files will not be downloaded
if they are found in \code{dir}. Set this to \code{TRUE} to download and overwrite
existing files.}
}
\value{
A tibble with the columns \code{compound_ids}, \code{pathway_ids} (if used),
and \code{mol_paths} (paths to downloaded .mol files).
}
\description{
Downloads mol files corresponding to individual compounds or compounds in a
pathway from KEGG.
}
\note{
For additional functionality for interacting with KEGG, try the
\code{KEGGREST} package, which this function was inspired by.
}
\examples{
\dontrun{
get_mol_kegg(compound_ids = c("C16181", "C06074"), dir = tempdir())
get_mol_kegg(pathway_ids = "map00253", dir = tempdir())
}
}
