% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mol_kegg.R
\name{get_mol_kegg}
\alias{get_mol_kegg}
\title{Download compound .mol files from KEGG}
\usage{
get_mol_kegg(compound_ids, pathway_ids, dir)
}
\arguments{
\item{compound_ids}{character vector of KEGG compound IDs---5 digits
prepended with a "C".}

\item{pathway_ids}{character vector of KEGG pathway IDs---5 digits prepended
with "map".}

\item{dir}{path to a folder to save mol files in. Folder will be created if
it does not already exist}
}
\value{
a tibble with the columns \code{compound_ids}, \code{pathway_ids} (if used),
and \code{mol_paths} (paths to downloaded .mol files)
}
\description{
Downloads mol files corresponding to individual compounds or compounds in a
pathway from KEGG.
}
\examples{
\dontrun{
get_mol_kegg(compound_ids = c("C16181", "C06074"), dir = tempdir())
get_mol_kegg(pathway_ids = "map00253", dir = tempdir())
}
}
