% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_compound_mol.R
\name{save_compound_mol}
\alias{save_compound_mol}
\title{Download compound .mol file}
\usage{
save_compound_mol(
  compound_id = NULL,
  compound_formula = NULL,
  pathway_id = NULL,
  path = "data",
  redownload = FALSE
)
}
\arguments{
\item{compound_id}{A character string that is 5 digits prepended with a "C".}

\item{compound_formula}{A character string detailing a compound formula.}

\item{pathway_id}{An optional character string specifying KEGG pathway ID, in format of 5 digits prepended with "map".}

\item{path}{An optional parameter to set relative path to location to download data.}

\item{redownload}{Download file again even if it has already been downloaded at path.}
}
\value{
Downloaded .mol file for compound in path folder.
}
\description{
Using KEGG ID values for compound and pathway, download
and clean up corresponding mol file
}
\examples{
save_compound_mol(compound_id = "C16181")

}
