% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_compound_mol.R
\name{save_compound_mol}
\alias{save_compound_mol}
\title{Download compound .mol file}
\usage{
save_compound_mol(
  pathway_id,
  path,
  compound_id = NULL,
  compound_formula = NULL,
  redownload = FALSE
)
}
\arguments{
\item{pathway_id}{character string that is 5 digits prepended with "map"}

\item{path}{relative path to location to download data}

\item{compound_id}{character string that is 5 digits prepended with a "C"}

\item{compound_formula}{character string of compound formula}

\item{redownload}{download file again even if it has already been downloaded at path}
}
\value{
downloaded .mol file
}
\description{
Using KEGG ID values for compound and pathway, download
and clean up corresponding mol file
}
