% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_compound_mol.R
\name{save_compound_mol}
\alias{save_compound_mol}
\title{Download compound .mol file}
\usage{
save_compound_mol(compound_id, pathway_id)
}
\arguments{
\item{compound_id}{character string that is 5 digits prepended with a "C"}

\item{pathway_id}{character string that is 5 digits prepended with "map"}
}
\value{
downloaded .mol file
}
\description{
Using KEGG ID values for compound and pathway, download
and clean up corresponding mol file
}
