% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpol1.R
\name{simpol1}
\alias{simpol1}
\title{SIMPOL.1 method for calculating estimated volatility}
\usage{
simpol1(fx_groups, meredith = TRUE, temp_c = 20)
}
\arguments{
\item{fx_groups}{A data.frame or tibble with counts of functional groups
produced by \code{\link[=get_fx_groups]{get_fx_groups()}} (or manually, with the same column names).}

\item{meredith}{Logical; \code{FALSE}: use the original SIMPOL.1 method. \code{TRUE}:
use the modified version in Meredith et al. (2023).}

\item{temp_c}{Numeric; the temperature at which to calculate volatility
estimates in °C.}
}
\value{
The \code{fx_groups} tibble with the additional \code{log10_P} column.
}
\description{
Implements the SIMPOL.1 group contribution method for predicting liquid vapor
pressure of organic compounds as described in Pankow & Asher (2008) and a
modified version described in Meredith et al. (2023).  Users will not usually
use this function directly, but rather through \code{\link[=calc_vol]{calc_vol()}}.
}
\details{
The output includes a column for \code{log10_P} where
\eqn{\textrm{log}_{10} P_{\textrm{L},i}^\circ(T) = \sum_k\nu_{k,i}b_k(T)}, or
the sum of the counts of functional groups (\eqn{\nu_{k,i}}) times the
coefficients for each functional group (\eqn{b_K(T)}). Units are in log10
atmospheres.

The modified method in Meredith et al. (2023) adds the following additional
functional groups:

Using the same coefficient as nitrate
\itemize{
\item Phosphoric acid
\item Phosphoric ester
\item Sulfate
\item Sulfonate
\item Thiol
}

Using the same coefficient as ester
\itemize{
\item Carbothioester
}
}
\note{
The method described in Pankow & Asher (2008) was developed using data
between 0 °C and 120 °C, so extrapolating beyond that range is not
recommended.
}
\examples{
mol_path <- mol_example()[3]
sdf <- ChemmineR::read.SDFset(mol_path)
fx_groups <- get_fx_groups(sdf)
simpol1(fx_groups)
}
\references{
Meredith L, Ledford S, Riemer K, Geffre P, Graves K, Honeker L, LeBauer D,
Tfaily M, Krechmer J. 2023. Automating methods for estimating metabolite
volatility. Frontiers in Microbiology. \doi{10.3389/fmicb.2023.1267234}

Pankow, J.F., Asher, W.E. 2008. SIMPOL.1: a simple group
contribution method for predicting vapor pressures and enthalpies of
vaporization of multifunctional organic compounds. Atmos. Chem. Phys.
\doi{10.5194/acp-8-2773-2008}
}
\seealso{
\code{\link[=calc_vol]{calc_vol()}}
}
