% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpol1.R
\name{simpol1}
\alias{simpol1}
\title{SIMPOL.1 method for calculating estimated volatility}
\usage{
simpol1(fx_groups, meredith = TRUE)
}
\arguments{
\item{fx_groups}{a data.frame or tibble with counts of functional groups
produced by \code{\link[=get_fx_groups]{get_fx_groups()}} (or manually, with the same column names)}

\item{meredith}{logical; \code{FALSE}: use the original SIMPOL.1 method. \code{TRUE}:
use the modified version in Meredith et al. (2023).}
}
\value{
the \code{fx_groups} tibble with the additional \code{log10_P} column
}
\description{
Implements the SIMPOL.1 group contribution method for predicting liquid vapor
pressure of organic compounds as described in Pankow & Asher (2008) and a
modified version described in Meredith et al. (2023).  Users will not usually
use this function directly, but rather through \code{\link[=calc_vol]{calc_vol()}} which uses this
as the default (currently only) method.
}
\details{
The output includes a column for \code{log10_P} where
\eqn{\textrm{log}_{10} P_{\textrm{L},i}^\circ(T) = \sum_k\nu_{k,i}b_k(T)}, or
the sum of the counts of functional groups (\eqn{\nu_{k,i}}) times the
coefficients for each functional group (\eqn{b_K(T)}). Units are in log10
atmospheres.

The modified method in Meredith et al. (2023) adds the following additional
functional groups and coefficients:
\itemize{
\item Phosphoric acid (-2.23)
\item Phosphoric ester (-2.23)
\item Sulfate (-2.23)
\item Sulfonate (-2.23)
\item Thiol (-2.23)
\item Carbothioester (-1.20)
}
}
\note{
The method described in Pankow & Asher (2008) allows for
calculations of logP at different temperatures.  This implementation
currently only calculates values at 20ºC.
}
\examples{
mol_path <- mol_example("C16181.mol")
sdf <- ChemmineR::read.SDFset(mol_path)
fx_groups <- get_fx_groups(sdf)
simpol1(fx_groups)
}
\references{
Pankow, J.F., Asher, W.E., 2008. SIMPOL.1: a simple group
contribution method for predicting vapor pressures and enthalpies of
vaporization of multifunctional organic compounds. Atmos. Chem. Phys.
\url{https://doi.org/10.5194/acp-8-2773-2008}

Meredith, L.K., Riemer, K., Geffre, P., Honeker, L., Krechmer, J., Graves,
K., Tfaily, M., and Ledford, S.K. \emph{In review}. Automating methods for
estimating metabolite volatility.  Frontiers in Microbiology.
}
\seealso{
\code{\link[=calc_vol]{calc_vol()}}
}
