% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vol.R
\name{simpol1}
\alias{simpol1}
\title{SIMPOL.1 method for calculating estimated volatility}
\usage{
simpol1(fx_groups)
}
\arguments{
\item{fx_groups}{}
}
\value{
a \code{data.frame} object
}
\description{
Implements SIMPOL.1
}
