% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpol1.R
\name{simpol1}
\alias{simpol1}
\title{SIMPOL.1 method for calculating estimated volatility}
\usage{
simpol1(fx_groups)
}
\arguments{
\item{fx_groups}{a data.frame with counts of functional groups produced by
\code{get_fx_groups()}}
}
\value{
a \code{data.frame} object
}
\description{
Implements SIMPOL.1
}
