% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcalc-package.R
\docType{package}
\name{volcalc-package}
\alias{volcalc}
\alias{volcalc-package}
\title{volcalc: Calculate Volatility of Chemical Compounds}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate estimated relative volatility index values for organic compounds based on functional group contributions. Calculation uses the SIMPOL.1 method (Prankow and Asher, 2008) \doi{10.5194/acp-8-2773-2008} or modified SIMPOL.1 method as in Meredith et al. (2023) \doi{10.5194/acp-8-2773-2008}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://meredith-lab.github.io/volcalc/}
  \item Report bugs at \url{https://github.com/Meredith-Lab/volcalc/issues}
}

}
\author{
\strong{Maintainer}: Kristina Riemer \email{kristinariemer@arizona.edu} (\href{https://orcid.org/0000-0003-3802-3331}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eric R. Scott (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})
  \item Laura K. Meredith (\href{https://orcid.org/0000-0003-4244-4366}{ORCID})
}

Other contributors:
\itemize{
  \item S. Marshall Ledford [contributor]
  \item Tamás Stirling [contributor]
}

}
\keyword{internal}
